requirements.txt
acme==1.21.0
apt-clone==0.2.1
apturl==0.5.2
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
blinker==1.6.2
Brlapi==0.8.3
certbot==1.21.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
distro==1.7.0
exceptiongroup==1.1.3
executing==2.0.0
eyeD3==0.8.10
filelock==3.6.0
Flask==3.0.0
greenlet==3.0.0
grpcio==1.30.2
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
josepy==1.10.0
jupyter_client==8.4.0
jupyter_core==5.4.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
more-itertools==8.10.0
nemo-emblems==5.4.1
nest-asyncio==1.5.8
netaddr==0.8.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
paho-mqtt==1.6.1
PAM==0.4.2
parsedatetime==2.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
Pygments==2.16.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
python-xapp==2.2.2
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
scour==0.38.2
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.21
stack-data==0.6.3
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==3.0.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
